requirements.txt
flask==3.0.3
supabase==2.5.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
groq==0.9.0
